Add fake Google results page
This will be used in variants to explore how we might fall back to Google search results where JS is not enabled. This was previously explored in the ‘google’ branch but never merged back, and has become to difficult to rebase.

We can in the future just create branches and only have to change the search box to point to the google form.